Se cambiar orden de archivos de controladores y se ajustan métodos para sus respectivos controladores
