+++ src/main/java/com/eam/skillforge/capaPresentacion/controlador/ControladorAdministrador.java
@@ -1,35 +1,18 @@
-package com.eam.skillforge.capaNegocio.controlador;
+package com.eam.skillforge.capaPresentacion.controlador;
 
 import com.eam.skillforge.capaNegocio.dto.UsuarioDto;
-import com.eam.skillforge.capaPersistencia.entidad.Usuario;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.apache.coyote.Response;
-import org.springframework.http.HttpRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
-import java.util.List;
-
 @RestController
 @RequestMapping("/usuarios")
 @Tag(name = "Usuarios")
-public class ControladorUsuario {
-
-    @GetMapping("/")
-    @Operation(summary = "Listar usuarios")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Lista de usuarios"),
-            @ApiResponse(responseCode = "401", description = "Este usuario no es permitido")
-    })
-    public ResponseEntity<ArrayList<UsuarioDto>> getUsuarios() {
-        return new ResponseEntity<>(new ArrayList<UsuarioDto>(), HttpStatus.OK);
-    }
-
+public class ControladorAdministrador {
     @PostMapping("")
     @Operation(summary = "Crear usuario")
     @ApiResponses(value = {
