+++ src/main/java/com/eam/skillforge/capaPresentacion/controlador/MedallaControlador.java
@@ -26,6 +26,11 @@ import org.springframework.web.bind.annotation.*;
 @CrossOrigin(origins = "*")
 public class MedallaControlador {
     private final MedallaServicio medallaServicio;
+
+    public MedallaControlador(MedallaServicio medallaServicio) {
+        this.medallaServicio = medallaServicio;
+    }
+
     /**
      * Crear una medalla nueva
      */
